ES-5578 Fixed minimumRenewals and maximumRenewals type

+++ src/main/kotlin/com/procurement/access/application/model/params/DivideLotParams.kt
@@ -101,8 +101,8 @@ data class DivideLotParams(
 
             data class Renewal(
                 val description: String?,
-                val minimumRenewals: Long?,
-                val maximumRenewals: Long?,
+                val minimumRenewals: Int?,
+                val maximumRenewals: Int?,
                 val period: Period?
             ) {
                 data class Period(
